feat: 更新模型配置和脚本，优化LoRA训练流程
refactor: 重构SVDLLM模块，添加新版本V1实现

fix: 修正.gitignore文件，添加.lora目录忽略

chore: 新增0.profile_time.sh脚本用于GPU监控

perf: 将模型精度从fp16改为bf16以提升性能

docs: 更新任务列表，移除boolq评估任务

+++ utils/LoRA.py
@@ -44,7 +44,7 @@ def apply_lora(model, tokenizer, batch_size=64, micro_batch_size=4, cutoff_len=2
     prompter = ZeroPrompter()
 
     if device == 'cuda':
-        model.half()
+        model.to(torch.bfloat16)
 
     tokenizer.pad_token_id = 0
     tokenizer.padding_side = "left"
@@ -155,7 +155,7 @@ def apply_lora(model, tokenizer, batch_size=64, micro_batch_size=4, cutoff_len=2
             warmup_steps=100,
             num_train_epochs=num_epochs,
             learning_rate=learning_rate,
-            fp16=True,
+            bf16=True,
             logging_steps=10,
             logging_first_step=True,
             optim="adamw_torch",
@@ -202,7 +202,7 @@ def main(args):
         prompter = ZeroPrompter()
 
     if device == 'cuda':
-        model.half()
+        model.to(torch.bfloat16)
 
     tokenizer.pad_token_id = 0
     tokenizer.padding_side = "left"
